Added manage job listings page - added authorization to allow right author to manage their job listings

+++ app/Http/Controllers/ListingController.php
@@ -59,6 +59,8 @@ class ListingController extends Controller
 
     public function update(Listing $listing, Request $request)
     {
+        if($listing->user_id !== auth()->id()) abort(403, 'Unauthorized Action');
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -80,8 +82,16 @@ class ListingController extends Controller
 
     public function destroy(Listing $listing)
     {
+        if($listing->user_id !== auth()->id()) abort(403, 'Unauthorized Action');
+
         $listing->delete();
 
         return redirect(@route('home'))->with('message', 'Listing successfully deleted');
+    }
+
+    public function manage()
+    {
+        $listings = auth()->user()?->listings()->latest()->get();
+        return view('listings.manage', compact('listings'));
     }
 }
